gui from unstable branch

+++ gui/objects/image_viewer.py
@@ -1,10 +1,7 @@
 
-from os import environ
-environ['PYGAME_HIDE_SUPPORT_PROMPT'] = '1'  # noqa
-import pygame
-
 import os
 
+import pygame  # GUI Framework
 from PIL import Image as im  # Import library that reads image array
 
 from cuticle_analysis import const
@@ -25,8 +22,8 @@ class ImageViewer():
         self.bkg_color = (50, 50, 50)
         self.bkg_obj = pygame.Rect(position[0], position[1], size[0], size[1])
         self.size = size
-        self.increment_id = False
-        self.decrement_id = False
+        self.increment_flag = False
+        self.decrement_flag = False
 
     def __scale_image__(self):
         """
@@ -65,35 +62,43 @@ class ImageViewer():
 
     def __get_increment_flag__(self):
         "Returns the increment id flag for the main thread to increment image id."
-        return self.increment_id
+        return self.increment_flag
 
     def __get_decrement_flag__(self):
         "Returns the increment id flag for the main thread to decrement image id."
-        return self.decrement_id
+        return self.decrement_flag
 
-    def __set_increment_flag__(self, new_flag):
+    def __increment_image__(self):
         "Sets the increment id flag for the main thread to increment image id."
-        self.increment_id = new_flag
+        if self.id < 2876:
+            self.id = self.id + 1
 
-    def __set_decrement_flag__(self, new_flag):
+    def __decrement_image__(self):
         "Sets the decrement id flag for the main thread to decrement image id."
-        self.decrement_id = new_flag
+        if self.id > 1:
+            self.id -= 1
 
-    def __update_image__(self, id):
+    def __correct_color__(self):
+        """Corrects the color of the image by swapping blue and red color values."""
+        print(str(self.image_arr[0][0]))
+
+    def __update_image__(self, id=None):
         """
             Updates the image viewer to current image, which is an image cache.
             Post-condition: Ant image is shown in image viewer. If file not found,
             The cache image is deleted and the image viewer shows error.
         """
-        if id >= 1 and id <= 1773:
-            self.id = id
+        #IMPLEMENT FUNCTION THAT OBTAINS THE MAXIMUM IMAGE ID.
+        if (id >= 1 and id <= 2876) or id is None:
+            if id is not None:
+                self.id = id
             try:
                 self.image_arr = im.fromarray(
                     self.image_list.get_image(self.id))
+                self.__correct_color__()
                 self.image_arr.save(self.cache_path)
                 return 1
             except Exception as e:
-                print(e)
                 self.image_arr = None
                 self.__delete_img_cache__()
                 return -1
@@ -101,5 +106,14 @@ class ImageViewer():
 
     def __delete_img_cache__(self):
         "Deletes the image cache file upon program exit."
-        os.remove(self.cache_path)
+        try:
+            if os.path.exists(self.cache_path) is True:
+                os.remove(self.cache_path)
+        except Exception as e:
+            print("Cache deleted.")
         self.image = None
+
+    
+    def get_image_id(self):
+        """Returns the image id."""
+        return self.id
